+++ parallel/tests_parallel/kernel/test_p_timeline.py
@@ -6,9 +6,6 @@ from sequence.kernel.event import Event
 
 from psequence.p_timeline import ParallelTimeline
 
-rank = MPI.COMM_WORLD.Get_rank()
-size = MPI.COMM_WORLD.Get_size()
-
 
 class FakeEntity(Entity):
     def __init__(self, name, tl):
@@ -22,7 +19,7 @@ class FakeEntity(Entity):
         self.counter += 1
 
 
-def build_env(lookahead):
+def build_env(lookahead, rank, size):
     tl = ParallelTimeline(lookahead)
 
     entity = FakeEntity(str(rank), tl)
@@ -34,15 +31,22 @@ def build_env(lookahead):
 
 @pytest.mark.mpi
 def test_p_timeline_schedule_local_events():
-    tl, entity = build_env(10)
+    rank = MPI.COMM_WORLD.Get_rank()
+    size = MPI.COMM_WORLD.Get_size()
+
+    tl, entity = build_env(10, rank, size)
     assert len(tl.events) == 0
     event = Event(20, Process(entity, "add", []))
     tl.schedule(event)
     assert len(tl.events) == 1
 
 
+@pytest.mark.mpi
 def test_p_timeline_schedule_remote_events():
-    tl, entity = build_env(10)
+    rank = MPI.COMM_WORLD.Get_rank()
+    size = MPI.COMM_WORLD.Get_size()
+
+    tl, entity = build_env(10, rank, size)
     assert len(tl.events) == 0
     foreign_rank = (int(entity.name) + 1) % size
     foreign_entity_name = str(foreign_rank)
@@ -52,8 +56,12 @@ def test_p_timeline_schedule_remote_events():
     assert len(tl.event_buffer[foreign_rank]) == 1
 
 
+@pytest.mark.mpi
 def test_p_timeline_run():
-    tl, entity = build_env(10)
+    rank = MPI.COMM_WORLD.Get_rank()
+    size = MPI.COMM_WORLD.Get_size()
+
+    tl, entity = build_env(10, rank, size)
     assert len(tl.events) == 0
     foreign_rank = (int(entity.name) + 1) % size
     foreign_entity_name = str(foreign_rank)