+++ tests/kernel/test_quantum_manager.py
@@ -1,4 +1,5 @@
 import numpy as np
+from scipy.linalg import fractional_matrix_power
 import math
 
 from sequence.kernel.quantum_manager import *
@@ -341,3 +342,214 @@ def test_qmanager__measure_density():
 
     assert abs((len(meas_0) / NUM_TESTS) - 0.5) < 0.1
 
+
+def test_qmanager__prepare_state_fock():
+    qm = QuantumManagerDensityFock(truncation=2)
+    state_0 = [1, 0, 0]
+    state_1 = [0, 1, 0]
+    combined = np.kron(state_0, state_1)
+    desired_state = np.outer(combined, combined.conj())
+
+    # test single state
+    key_0 = qm.new()
+    key_1 = qm.new()
+    qm.set([key_0, key_1], combined)
+
+    new_state, all_keys = qm._prepare_state([key_0, key_1])
+    assert np.all(new_state == np.array(desired_state))
+    assert np.all(np.array(all_keys) == np.array([key_0, key_1]))
+
+    # test disjoint state
+    key_0 = qm.new(state_0)
+    key_1 = qm.new(state_1)
+
+    new_state, all_keys = qm._prepare_state([key_0, key_1])
+    assert np.all(new_state == np.array(desired_state))
+    assert np.all(np.array(all_keys) == np.array([key_0, key_1]))
+
+    # test state that needs swapping
+    key_1 = qm.new()
+    key_0 = qm.new()
+    combined_2 = np.kron(state_1, state_0)
+    qm.set([key_1, key_0], combined_2)
+
+    new_state, all_keys = qm._prepare_state([key_0, key_1])
+    assert np.all(new_state == np.array(desired_state))
+    assert np.all(np.array(all_keys) == np.array([key_0, key_1]))
+
+    # test larger state that needs swapping
+    key_0 = qm.new()
+    key_1 = qm.new()
+    key_2 = qm.new()
+    combined_3 = np.kron(np.kron(state_0, state_0), state_1)
+    qm.set([key_0, key_1, key_2], combined_3)
+
+    combined_correct = np.kron(np.kron(state_0, state_1), state_0)
+    desired_state_alt = np.outer(combined_correct, combined_correct.conj())
+    new_state, all_keys = qm._prepare_state([key_0, key_2])
+    assert np.all(new_state == np.array(desired_state_alt))
+    assert np.all(np.array(all_keys) == np.array([key_0, key_2, key_1]))
+
+
+def test_qmanager_build_ladder_fock():
+    TRUNCATION = 2
+
+    qm = QuantumManagerDensityFock(truncation=TRUNCATION)
+    create, destroy = qm.build_ladder()
+
+    # check dims
+    assert len(create.shape) == 2
+    assert len(destroy.shape) == 2
+    assert create.shape[0] == create.shape[1] == TRUNCATION + 1
+    assert destroy.shape[0] == destroy.shape[1] == TRUNCATION + 1
+
+    # check properties
+    state = [0] * (TRUNCATION + 1)
+    state[-1] = 1
+    state = array(state)
+    state_create = create @ state
+    assert all(state_create == 0)
+
+    state = [0] * (TRUNCATION + 1)
+    state[0] = 1
+    state = array(state)
+    state_destroy = destroy @ state
+    assert all(state_destroy == 0)
+
+    for i in range(TRUNCATION):
+        state = [0] * (TRUNCATION + 1)
+        state[i] = 1
+        state = array(state)
+
+        state_create = create @ state
+        assert state_create[i+1] == sqrt(i+1)
+        assert sum(state_create) == sqrt(i+1)
+
+    for i in range(TRUNCATION):
+        state = [0] * (TRUNCATION + 1)
+        state[i+1] = 1
+        state = array(state)
+
+        state_destroy = destroy @ state
+        assert state_destroy[i] == sqrt(i+1)
+        assert sum(state_destroy) == sqrt(i+1)
+
+
+def test_qmanager_apply_operator_fock():
+    TRUNCATION = 2
+
+    qm = QuantumManagerDensityFock(truncation=2)
+    create, destroy = qm.build_ladder()
+    excited = np.zeros((TRUNCATION + 1, TRUNCATION + 1))
+    excited[1, 1] = 1
+
+    # single state
+    key = qm.new()
+    qm.apply_operator(create, [key])
+    new_state = qm.get(key)
+    assert np.all(new_state.state == excited)
+
+    # compound state
+    key1 = qm.new()
+    key2 = qm.new()
+    oper = np.eye((TRUNCATION + 1) ** 2)
+    qm.apply_operator(oper, [key1, key2])
+    qm.apply_operator(create, [key1])
+    new_state = qm.get(key1)
+
+    desired2 = np.zeros((TRUNCATION + 1, TRUNCATION + 1))
+    desired2[0, 0] = 1
+    desired = np.kron(excited, desired2)
+    assert np.all(new_state.state == desired)
+
+    # comound operator
+    key1 = qm.new()
+    key2 = qm.new()
+    oper = np.kron(create, create)
+    qm.apply_operator(oper, [key1, key2])
+    new_state = qm.get(key1)
+
+    desired = np.kron(excited, excited)
+    assert np.all(new_state.state == desired)
+
+
+def test_qmanager_measure_fock():
+    NUM_TESTS = 1000
+    TRUNCATION = 2
+
+    qm = QuantumManagerDensityFock(truncation=2)
+    create, destroy = qm.build_ladder()
+    series_elem_list = [((-1) ** i) * fractional_matrix_power(create, i + 1).dot(
+        fractional_matrix_power(destroy, i + 1)) / math.factorial(i + 1) for i in range(TRUNCATION)]
+    povm1 = sum(series_elem_list)
+    povm0 = eye(TRUNCATION + 1) - povm1
+
+    # single state
+    meas_0 = []
+    meas_1 = []
+    state_single = [math.sqrt(1 / 2), math.sqrt(1 / 2), 0]
+    state = np.outer(state_single, state_single)
+    for _ in range(NUM_TESTS):
+        key = qm.new(state)
+        samp = np.random.random()
+        res = qm.measure([key], [povm0, povm1], samp)
+        if res:
+            meas_1.append(key)
+        else:
+            meas_0.append(key)
+
+    assert abs((len(meas_0) / NUM_TESTS) - 0.5) < 0.1
+    for key in meas_0:
+        assert qm.states[key] is None
+    for key in meas_1:
+        assert qm.states[key] is None
+
+    # single state in multi-qubit system
+    meas_0 = []
+    meas_1 = []
+    for _ in range(NUM_TESTS):
+        key1 = qm.new(state)
+        key2 = qm.new()
+        samp = np.random.random()
+        # compound
+        oper = np.identity((TRUNCATION + 1) ** 2)
+        qm.apply_operator(oper, [key1, key2])
+        res = qm.measure([key1], [povm0, povm1], samp)
+        if res:
+            meas_1.append(key2)
+        else:
+            meas_0.append(key2)
+
+    assert abs((len(meas_0) / NUM_TESTS) - 0.5) < 0.1
+
+    desired = [0] * (TRUNCATION + 1)
+    desired[0] = 1
+    desired_density = np.outer(desired, desired)
+    for key in meas_0:
+        assert np.all(qm.get(key).state == desired_density)
+    for key in meas_1:
+        assert np.all(qm.get(key).state == desired_density)
+
+    # multiple state
+    povm00 = np.kron(povm0, povm0)
+    povm01 = np.kron(povm0, povm1)
+    povm10 = np.kron(povm1, povm0)
+    povm11 = np.kron(povm1, povm1)
+    meas_0 = []
+    meas_2 = []
+    for _ in range(NUM_TESTS):
+        key1 = qm.new(state)
+        key2 = qm.new()
+        samp = np.random.random()
+        # compound
+        oper = np.identity((TRUNCATION + 1) ** 2)
+        qm.apply_operator(oper, [key1, key2])
+        res = qm.measure([key1, key2], [povm00, povm01, povm10, povm11], samp)
+        if res == 2:
+            meas_2.append(key1)
+        elif res == 0:
+            meas_0.append(key1)
+        else:
+            raise Exception()
+
+    assert abs((len(meas_0) / NUM_TESTS) - 0.5) < 0.1